cast int to remove warning

+++ src/lib/pcg-c-basic-master/pcg32x2-demo.c
@@ -152,7 +152,7 @@ int main(int argc, char** argv)
         for (i = 0; i < 6; ++i) {
             if (i > 0 && i % 3 == 0)
                 printf("\n\t");
-            printf(" 0x%016llx", pcg32x2_random_r(&rng));
+            printf(" 0x%016llx", (long long unsigned int) pcg32x2_random_r(&rng));
         }
         printf("\n");
